VTryon_Updated/requirements.txt
# --- Frontend & UI (Your requested versions) ---
streamlit>=1.30.0
Pillow>=10.0.0
python-dotenv>=1.0.0
google-generativeai>=0.5.0

# --- Backend & Server (Added for FastAPI) ---
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
requests>=2.31.0

# --- Modern GenAI Client ---
google-genai>=0.8.0